% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adrr.R
\name{adrr_flights_tidy}
\alias{adrr_flights_tidy}
\title{Extract ADRR flight list in an interval of dates}
\usage{
adrr_flights_tidy(conn = NULL, wef, til)
}
\arguments{
\item{conn}{Database connection or instantiate the default one.}

\item{wef}{\strong{W}ith \strong{EF}fect date (included) at Zulu time
in a format recognized by \code{\link[lubridate:as_date]{lubridate::as_datetime()}}}

\item{til}{un\strong{TIL}l date (excluded) at Zulu time
in a format recognized by \code{\link[lubridate:as_date]{lubridate::as_datetime()}}}
}
\value{
A \code{\link[dbplyr:tbl.src_dbi]{dbplyr::tbl_dbi()}} with the following columns (as defined in the
Aviation Data Repository for Research manual):
\subsection{Flight details}{
\itemize{
\item \verb{ECTRL ID}: flight unique id.
\item \code{ADEP}: \href{https://observablehq.com/@openaviation/airports}{ICAO code} of the
\item \verb{ADEP Latitude}: \code{ADEP} latitude (decimal degrees North)
\item \verb{ADEP Longitude}: \code{ADEP} longitude (decimal degrees East)
\strong{A}erodrome of \strong{DEP}arture
\item \code{ADES}: \href{https://observablehq.com/@openaviation/airports}{ICAO code} of the
\strong{A}erodrome of \strong{DES}tination
\item \verb{ADES Latitude}: \code{ADES} latitude (decimal degrees North)
\item \verb{ADES Longitude}: \code{ADES} longitude (decimal degrees East)
\item \verb{FILED OFF BLOCK TIME}: (Last) filed off-block time
\item \verb{FILED ARRIVAL TIME}: (Last) filed arrival time
\item \verb{ACTUAL OFF BLOCK TIME}: actual off-block time for flown (M3) trajectory
\item \verb{ACTUAL ARRIVAL TIME}: arrival time for flown (M3) trajectory
\item \verb{AC Type}: the \href{https://www.icao.int/publications/doc8643/pages/search.aspx}{ICAO code for the aircraft type},
for example A30B for an Airbus A-300B2-200.
\item \verb{AC Operator}: the \href{https://en.wikipedia.org/wiki/List_of_airline_codes}{ICAO Airline Designator},
i.e. \code{OAL} for \code{Olympic}
\item \verb{AC Registration}: the \href{https://en.wikipedia.org/wiki/Aircraft_registration}{aircraft registration}
(with spaces, dashes, ... stripped), e.g. GEUUU
\item \verb{ICAO Flight Type}: flight type (see
\href{https://www.skybrary.aero/articles/flight-plan-completion}{FPL Item 8}).
Possible values:
\itemize{
\item \code{S} for scheduled air service
\item \code{N} for non-scheduled air service
}
\item \verb{STATFOR Market Segment}: market segment type as defined on the
\href{https://www.eurocontrol.int/publication/market-segment-rules}{Market Segment Rules},
it can be:
\itemize{
\item “Mainline”
\item “Regional”
\item “Low-Cost”
\item “Business Aviation”
\item “All-Cargo”
\item “Charter” (Non-Scheduled)
\item “Military”
\item “Other”
\item "Not classified"
}
\item \verb{Requested FL}: requested cruising flight level (FL)
\item \verb{Actual Distance Flown (nm)}: route length (in Nautical Miles) for for flown
(M3) trajectory
}
}
}
\description{
Extract ADRR flight list in an interval of dates
}
\section{Note}{
You need to either provide a connection \code{conn} that has access to
\code{SWH_FCT.DIM_FLIGHT_TYPE_RULE} (for \code{FLT_RULES}),
\code{PRUDEV.V_COVID_DIM_AO} (for aircraft and aircraft group info) and
\code{SWH_FCT.V_FAC_FLIGHT_MS} (for market segment info) or go with the default
which uses PRU_READ to establish a \code{\link[=db_connection]{db_connection()}}.
}

\examples{
\dontrun{
withr::local_envvar(c(TZ = "UTC", ORA_SDTZ = "UTC", NLS_LANG = ".AL32UTF8"))
conn <- withr::local_db_connection(db_connection(schema = "PRU_READ"))
wef <- "2025-01-01"
til <- "2025-01-04"
adrr_flights_tidy(conn, wef, til)
}
}
