% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{aircraft_model}
\alias{aircraft_model}
\title{ICAO's Manufacturer codes}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 7260 rows and 4 columns.
}
\usage{
aircraft_model
}
\description{
A data frame with the following fields
\describe{
\item{model_full_name}{the model full name, e.g. "A-320neo".}
\item{manufacturer_code}{the manufacturer's code, e.g. "AIRBUS".}
\item{designator}{the model's designator, e.g. "A20N"}
\item{last_updated}{the date when the data have been last updated, e.g. "2023-05-19".}
}
}
\keyword{datasets}
