% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/airlines.R
\name{airlines_tbl}
\alias{airlines_tbl}
\title{Return a reference to the Airlines table}
\usage{
airlines_tbl(conn = NULL)
}
\arguments{
\item{conn}{Database connection or instantiate the default one.}
}
\value{
a \code{\link[dbplyr:tbl.src_dbi]{dbplyr::tbl_dbi()}} referencing the Oracle table for airlines.
}
\description{
The returned \code{\link[dbplyr:tbl.src_dbi]{dbplyr::tbl_dbi()}} is referencing the airlines table in PRISME.
You can use \code{dplyr}/\code{dbplyr} verbs to filter, join, ... with other
datasets.
}
\section{Note}{
You need to either provide a connection \code{conn}
that has access to \code{PRU_DEV.V_COVID_DIM_AO} or go with the
default which uses PRU_DEV to establish a \code{\link[=db_connection]{db_connection()}}.
}

\examples{
\dontrun{
arl <- airlines_tbl()
# other operations on arl, i.e. filtering,
# followed by a collect() to retrieve the concrete data frame
arl_filtered <- arl |>
   dplyr::filter(AO_ISO_CTRY_CODE == "IT") |>
   collect()

# NOTE: you can reuse the connection for other API calls
conn <- arl$src$con

# other ops requiring conn
# ...

# IMPORTANT: close the DB connection
DBI::dbDisconnect(conn)
}
}
