% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{airlines_tidy}
\alias{airlines_tidy}
\title{Airline info including group affiliation}
\usage{
airlines_tidy(con = NULL)
}
\arguments{
\item{con}{Optional connection to the PRU_DEV schema.}
}
\value{
A data frame with airline info such as:
* \code{operator_code}: the airline's ICAO code, i.e. OAL,
* \code{operator_name}: the airline's name, i.e. Olympic,
* \code{operator_group}: the airline's affiliation group, i.e. AEGEAN Group,
* \code{iso2c}: the ISO2C code of the airline's country, i.e. GR,
* \code{eu}: whether the airlines is in a EUROCONTROL's Member State (plus Kosovo), i.e. TRUE.
}
\description{
The
}
\examples{
\dontrun{
airlines_tidy()
}
}
