% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other.R
\name{airports_oa}
\alias{airports_oa}
\title{retrieve latest airport list from OurAirports}
\usage{
airports_oa()
}
\value{
a data frame
}
\description{
retrieve latest airport list from OurAirports
}
\examples{
\dontrun{
apts <- airports_oa()
}
}
