% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{airspace_profiles_tidy}
\alias{airspace_profiles_tidy}
\title{Provide all airspace profile segments intersecting an interval of interest}
\usage{
airspace_profiles_tidy(wef, til, airspace = "FIR", profile = "CTFM")
}
\arguments{
\item{wef}{\strong{W}ith \strong{EF}fect date (included) at Zulu time
in a format recognized by \code{\link[=as_datetime]{as_datetime()}}}

\item{til}{\strong{TIL}l date (excluded) at Zulu time
in a format recognized by \code{\link[=as_datetime]{as_datetime()}}}

\item{airspace}{the type of airspace (default: 'FIR'), one of:
\itemize{
\item 'FIR' (\href{https://observablehq.com/@openaviation/flight-information-regions}{Flight Information Region})
\item 'NAS' (National Airspace)
\item 'AUA' (ATC Unit Airspace)
\item 'ES' (Elementary Sector)
}}

\item{profile}{the model of the trajectory profil (default: 'CTFM'), one of:
\itemize{
\item "FTFM"
\item "CTFM"
\item "RTFM"
\item "CPF"
}}
}
\value{
a \link{tbl} with the following columns
\itemize{
\item ID: the so called SAM ID, used internally by PRISME
\item SEQ_ID: the sequence number of the segment for the relevant airspace profile
\item ENTRY_TIME: the time of entry into the relevant airspace
\item ENTRY_LON:  the longitude of entry into the relevant airspace
\item ENTRY_LAT: the latitude of entry into the relevant airspace
\item ENTRY_FL: the flight level of entry into the relevant airspace
\item EXIT_TIME: the time of exit out of the relevant airspace
\item EXIT_LON: the longitude of exit out of the relevant airspace
\item EXIT_LAT: the latitude of exit out of the relevant airspace
\item EXIT_FL: the flight level of exit out of the relevant airspace
\item AIRSPACE_ID: the airspace ID
\item AIRSPACE_TYPE: the airspace type as per \code{airspace} input parameter
\item MODEL_TYPE: the trajectory model as per \code{profile} input parameter
}
}
\description{
Provide all airspace profile segments intersecting an interval of interest
}
\examples{
\dontrun{
asp_profs <- airspace_profiles_tidy(wef = "2023-01-01", til = "2023-04-01")
}
}
