% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/airport_operator_data_flow.R
\name{aodf_tidy}
\alias{aodf_tidy}
\title{Extract a clean airport operator data flow list in an interval}
\usage{
aodf_tidy(conn = NULL, wef, til)
}
\arguments{
\item{conn}{Database connection or instantiate the default one.}

\item{wef}{\strong{W}ith \strong{EF}fect date (included) at Zulu time
in a format recognized by \code{\link[lubridate:as_date]{lubridate::as_datetime()}}}

\item{til}{un\strong{TIL}l date (excluded) at Zulu time
in a format recognized by \code{\link[lubridate:as_date]{lubridate::as_datetime()}}}
}
\value{
A \code{\link[dplyr:tbl]{dplyr::tbl()}} with the following columns:
\itemize{
\item APDS_ID: the airport operator dataflow unique record id.
\item ID: the so called \verb{SAM ID}, used internally by PRISME
\item AP_C_FLTID: flight identifier (aource Airport)
\item AP_C_FLTRUL: which sets of regulations the flight is operated under.
Possible values are:
}
\itemize{
\item \code{IFR} for IFR
\item \code{VFR} for VFR
\item \code{NA} if unknown
}
\itemize{
\item AP_C_REG: the \href{https://en.wikipedia.org/wiki/Aircraft_registration}{aircraft registration}
(with spaces, dashes, ... stripped), e.g. GEUUU.
\item ADEP_ICAO: (\href{https://observablehq.com/@openaviation/airports}{ICAO code} of the)
\strong{A}erodrome of \strong{DEP}arture (source airport).
\item ADES_ICAO: (\href{https://observablehq.com/@openaviation/airports}{ICAO code} of the)
\strong{A}erodrome of \strong{DES}tination  (source airport).
\item SRC_PHASE: flight phase. \code{DEP}=departure, \code{ARR}=arrival.
\item MVT_TIME_UTC: (best available) movement time
(takeoff if \code{SRC_PHASE} = \code{DEP}, landing if \code{SRC_PHASE} = \code{ARR}).
\item BLOCK_TIME_UTC: Block time
(off-block if \code{SRC_PHASE} = \code{DEP}, in-block if \code{SRC_PHASE} = \code{ARR}).
\item SCHED_TIME_UTC: scheduled time
(of departure if \code{SRC_PHASE} = \code{DEP}, of arrival if \code{SRC_PHASE} = \code{ARR}; source airport).
\item ARCTYP: (best available) the \href{https://www.icao.int/publications/doc8643/pages/search.aspx}{ICAO code for the aircraft type},
for example A21N for Airbus A321neo.
\item AP_C_RWY: Runway ID (of departure if \code{SRC_PHASE} = \code{DEP}, of arrival if \code{SRC_PHASE} = \code{ARR}; source airport).
\item AP_C_STND: Stand ID (of departure if \code{SRC_PHASE} = \code{DEP}, of arrival if \code{SRC_PHASE} = \code{ARR}; source airport).
\item C40_CROSS_TIME: time of first (last) crossing at 40 NM from ARP for departure (arrival).
\item C40_CROSS_LAT: latitude of first (last) crossing at 40 NM from ARP for departure (arrival).
\item C40_CROSS_LON: longitude of first (last) crossing at 40 NM from ARP for departure (arrival).
\item C40_CROSS_FL: flight level of first (last) crossing at 40 NM from ARP for departure (arrival).
\item C40_BEARING: bearing of first (last) crossing at 40 NM from ARP for departure (arrival).
\item C100_CROSS_TIME: time of first (last) crossing at 100 NM from ARP for departure (arrival).
\item C100_CROSS_LAT: latitude of first (last) crossing at 100 NM from ARP for departure (arrival).
\item C100_CROSS_LON: longitude of first (last) crossing at 100 NM from ARP for departure (arrival).
\item C100_CROSS_FL: flight level of first (last) crossing at 100 NM from ARP for departure (arrival).
\item C100_BEARING: bearing of first (last) crossing at 100 NM from ARP for departure (arrival).
}
}
\description{
The returned \code{\link[dplyr:tbl]{dplyr::tbl()}} includes movements information in the
interval \verb{[wef, til)}.
\strong{NOTE}: it can only cover ONE month at a time
}
\section{Note}{
You need to either provide a connection \code{conn} that has access to \code{SWH_FCT.FAC_APDS_FLIGHT_IR691},
or go with the default which uses PRU_ATMAP to establish a \code{\link[=db_connection]{db_connection()}}.
}

\examples{
\dontrun{
my_aodf <- aodf_tidy(wef = "2023-01-01", til = "2023-01-02")
# ...
DBI::dbDisconnect(my_aodf$src$con)
}
}
