% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/airport_operator_data_flow.R
\name{apdf_tbl}
\alias{apdf_tbl}
\title{Return a reference to the Airport Operator Data Flow table}
\usage{
apdf_tbl(conn = NULL)
}
\arguments{
\item{conn}{Database connection or instantiate the default one.}
}
\value{
a \code{\link[dbplyr:tbl.src_dbi]{dbplyr::tbl_dbi()}} referencing the Oracle table for airport
operator data flow.
}
\description{
The returned \code{\link[dbplyr:tbl.src_dbi]{dbplyr::tbl_dbi()}} is referencing the airport operator data
flow table in PRISME.
You can use \code{dplyr}/\code{dbplyr} verbs to filter, join, ... with other
datasets.
}
\section{Note}{
You need to either provide a connection \code{conn} that has access to
\code{SWH_FCT.FAC_APDS_FLIGHT_IR691} or go with the default which uses PRU_ATMAP
to establish a \code{\link[=db_connection]{db_connection()}}.
}

\examples{
\dontrun{
apdf <- apdf_tbl()
# ...
# IMPORTANT: close the DB connection when done with `aodf`
DBI::dbDisconnect(apdf$src$con)
}

}
