% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/airspaces.R
\name{es_sf}
\alias{es_sf}
\title{Retrieve Elementary Sector airspace geometries for a given AIRAC cycle}
\usage{
es_sf(conn = NULL, cfmu_airac)
}
\arguments{
\item{conn}{Database connection or instantiate the default one.}

\item{cfmu_airac}{the AIRAC cycle number (CFMU format), e.g. \code{"517"}}
}
\value{
An \code{sf} object with the following columns:
\itemize{
\item \code{AC_ID}: the AIRAC cycle id
\item \code{AV_AIRSPACE_ID}: the airspace identifier
\item \code{MIN_FLIGHT_LEVEL}: the minimum flight level
\item \code{MAX_FLIGHT_LEVEL}: the maximum flight level
\item \code{NAME}: the airspace name
\item \code{CODE}: the airspace code
\item \code{AIRSPACE_TYPE}: the airspace type (\code{ES})
\item \code{geometry}: the airspace polygon geometry
}
}
\description{
Returns ES (Elementary Sector) airspace geometries as an \code{sf} object for the
specified CFMU AIRAC cycle.
}
\section{Note}{
You need to either provide a connection \code{conn} that has access to
\code{ENV_SP.AIRSPACE}, \code{PRU_CFMU_ES} and \code{SWH_MAP} or go with the default
which uses PRU_READ to establish a \code{\link[=db_connection]{db_connection()}}.

The \code{sf} package is required but only suggested; install it if not available.
}

\examples{
\dontrun{
es <- es_sf(cfmu_airac = "517")

# write to GeoJSON
sf::st_write(es, "es-517.geojson", driver = "GeoJSON")

# if you re-use DB connections
conn <- eurocontrol::db_connection("PRU_READ")
es <- es_sf(conn = conn, cfmu_airac = "517")

# ... do something else with conn
# ...
# then manually close the connection to the DB
DBI::dbDisconnect(conn)
}
}
