% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eurocontrol-package.R
\docType{package}
\name{eurocontrol-package}
\alias{eurocontrol}
\alias{eurocontrol-package}
\title{eurocontrol: Helper functions for EUROCONTROL useRs}
\description{
The helper functions in this package are designed to make it easy, homogeneus and transparent to perform common tasks usually needed by data analysts and useRs in EUROCONTROL.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://eurocontrol.github.io/eurocontrol/}
  \item \url{https://github.com/eurocontrol/eurocontrol}
}

}
\author{
\strong{Maintainer}: Enrico Spinielli \email{enrico.spinielli@eurocontrol.int} (\href{https://orcid.org/0000-0001-8584-9131}{ORCID})

Authors:
\itemize{
  \item Enrico Spinielli \email{enrico.spinielli@eurocontrol.int} (\href{https://orcid.org/0000-0001-8584-9131}{ORCID})
  \item Quinten Goens \email{quinten.goens@eurocontrol.int} (\href{https://orcid.org/0000-0003-2187-1924}{ORCID})
}

Other contributors:
\itemize{
  \item EUROCONTROL [copyright holder, funder]
}

}
\keyword{internal}
