% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trajectories.R
\name{export_model_trajectory}
\alias{export_model_trajectory}
\title{Export point trajectories for different flight models.}
\usage{
export_model_trajectory(
  conn,
  wef,
  til,
  profile = "CTFM",
  bbox = NULL,
  lobt_buffer = c(before = 28, after = 24),
  timeover_buffer = NULL
)
}
\arguments{
\item{conn}{Database connection or instantiate the default one.}

\item{wef}{\strong{W}ith \strong{EF}fect date (included) at Zulu time
in a format recognized by \code{\link[lubridate:as_date]{lubridate::as_datetime()}}}

\item{til}{un\strong{TIL}l date (excluded) at Zulu time
in a format recognized by \code{\link[lubridate:as_date]{lubridate::as_datetime()}}}

\item{profile}{the \href{https://ansperformance.eu/definition/flight-models/}{model of the trajectory}
profile (default: 'CTFM'), one of:
\itemize{
\item 'FTFM', Filed Tactical Flight Model
\item 'RTFM', Regulated Tactical Flight Model
\item 'CTFM', Current Tactical Flight Model
\item 'CPF', Correlated Position reports for a Flight
\item 'DCT', Direct route
\item 'SCR', Shortest Constrained Route
\item 'SRR', Shortest RAD restrictions applied Route
\item 'SUR', Shortest Unconstrained Route
}}

\item{bbox}{(Optional) axis aligned bounding box (xmin, ymin, xmax, ymax)}

\item{lobt_buffer}{(Optional) (portion of) hours before/after
\code{wef} and \code{til} (before, after) used to query LOBT.
This is to cater for flights crossing \code{wef} and \code{til}.
For example \code{c(before = 24, after = 2.25)} allows to retrieve flights with LOBT
24H before \code{wef} and 1H15M after \code{til}.}

\item{timeover_buffer}{(Optional) (portion of) hours before (after) \code{wef} (\code{til}).
This is to cater for flights crossing \code{wef} and \code{til}.
For example \code{c(before = 2, after = 0.25)} allows to retrieve
points whose TIME_OVER is 2H before \code{wef} and 15M after \code{til}.}
}
\value{
a \code{tbl}
}
\description{
The returned \code{\link[dplyr:tbl]{dplyr::tbl()}} includes point profiles for flights
departing in the right-opened interval \verb{[wef, til)}.
}
\section{Note}{
You need to either provide a connection \code{conn} that has access to \code{SWH_FCT.DIM_FLIGHT_TYPE_RULE},
\code{PRUDEV.V_COVID_DIM_AO} and \code{SWH_FCT.FAC_FLIGHT} or go with the default which uses
PRU_DEV to establish a \code{\link[=db_connection]{db_connection()}}.
}

