% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trajectories.R
\name{flights_airspace_profiles_tidy}
\alias{flights_airspace_profiles_tidy}
\title{Extract the segments intersecting a set of airspace IDs in an interval
of time and complement them with flight information}
\usage{
flights_airspace_profiles_tidy(
  conn = NULL,
  wef,
  til,
  airspaces,
  airspace = "FIR",
  profile = "CTFM"
)
}
\arguments{
\item{conn}{Database connection or instantiate the default one.}

\item{wef}{\strong{W}ith \strong{EF}fect date (included) at Zulu time
in a format recognized by \code{\link[lubridate:as_datetime]{lubridate::as_datetime()}}}

\item{til}{un\strong{TIL}l date (excluded) at Zulu time
in a format recognized by \code{\link[lubridate:as_datetime]{lubridate::as_datetime()}}}

\item{airspaces}{list of airspace ids}

\item{airspace}{the type of airspace (default: 'FIR'), one of:
\itemize{
\item 'FIR' (\href{https://observablehq.com/@openaviation/flight-information-regions}{Flight Information Region})
\item 'NAS' (National Airspace)
\item 'AUA' (ATC Unit Airspace)
\item 'ES' (Elementary Sector)
}}

\item{profile}{the \href{https://ansperformance.eu/definition/flight-models/}{model of the trajectory}
profile (default: 'CTFM'), one of:
\itemize{
\item 'FTFM', Filed Tactical Flight Model
\item 'RTFM', Regulated Tactical Flight Model
\item 'CTFM', Current Tactical Flight Model
\item 'CPF', Correlated Position reports for a Flight
\item 'DCT', Direct route
\item 'SCR', Shortest Constrained Route
\item 'SRR', Shortest RAD restrictions applied Route
\item 'SUR', Shortest Unconstrained Route
}}
}
\value{
a \code{\link[dbplyr:tbl_dbi]{dbplyr::tbl_dbi()}} with the same columns as \code{\link[=flights_tidy]{flights_tidy()}}
plus the airspace id, \code{AIRSPACE_ID}, being crossed, one row per
crossed airspace.
}
\description{
The returned \code{\link[dbplyr:tbl_dbi]{dbplyr::tbl_dbi()}} includes scheduled and non-scheduled flights
whose airspace segments temporally intersecting the right-opened interval
\verb{[wef, til)}.
General aviation, State, military and sensitive flight are excluded.
}
\section{Note}{
You need to either provide a connection \code{conn} that has access to as noted in
\code{\link[=airspace_profile_tbl]{airspace_profile_tbl()}} and \code{\link[=flights_tidy]{flights_tidy()}} or go with the default which
uses PRU_READ to establish a \code{\link[=db_connection]{db_connection()}}.
}

\examples{
\dontrun{
withr::local_envvar(c(TZ = "UTC", ORA_SDTZ = "UTC", NLS_LANG = ".AL32UTF8"))
conn <- withr::local_db_connection(db_connection("PRU_READ"))
flights_airspace_profiles_tidy(
  conn = conn,
  wef = "2023-01-01",
  til = "2023-04-01",
  airspace = "ES",
  profile = "CTFM",
  airspaces = c("LYBALW2", "LYBATW2")
)
}
}
