% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{flights_tidy}
\alias{flights_tidy}
\title{Extract useful flights list from NM flight table is an interval}
\usage{
flights_tidy(wef, til, .cols = "USEFUL")
}
\arguments{
\item{wef}{With EFfect date (included) at Zulu time
in a format recognized by `lubridate::as_datetime()`}

\item{til}{TILl date (excluded) at Zulu time
in a format recognized by `lubridate::as_datetime()`}

\item{.cols}{either ALL, USEFUL or a character vector of column names [default: USEFUL]
USEFUL is a subset of typically used column names.}
}
\value{
a data frame with
* flight unique id,
* lobt, iobt, actualt OBT, day
* arrival datetime
* aircraft id (also CRCO and ACARS ones)
* aircraft registration (also from CRCO and ACARS)
* ...
* aircraft type
* aircraft operator
* aircraft group
* market segment
}
\description{
The returned `tbl` is referencing the flight table in PRISME and
including flights in the right-opened interval [wef, til).
You can use `dplyr`/`dbplyr` verbs to filter, join, ... with other
datasets.
NOTE: you need access to PRU_DEV
}
\examples{
\dontrun{
my_flts <- flights_tidy(wef = "2023-01-01", til = "2023-04-01")
}
}
