% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iata_season.R
\name{iata_season_for_date}
\alias{iata_season_for_date}
\title{Return the corresponding IATA season for a date}
\usage{
iata_season_for_date(date)
}
\arguments{
\item{date}{a date}
}
\value{
the name of the IATA season in the form \code{summer-yyyy}
}
\description{
Return the corresponding IATA season for a date
}
\examples{
\dontrun{
season_iata("2024-04-01")
}
}
