% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{member_states}
\alias{member_states}
\title{EUROCONTROL's Member States}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 45 rows and 7 columns.
}
\usage{
member_states
}
\description{
A data frame with the following fields
\describe{
\item{name}{the country name, e.g. "Italy"}
\item{iso3c}{the 3-letter ISO code, e.g. "ITA"}
\item{iso2c}{the 2-letter ISO code, e.g. "IT"}
\item{icao}{the 2-letter ICAO code, e.g. "LI"}
\item{iso3n}{the 3-digits ISO code, e.g. "380"}
\item{date}{the date of \code{status} code, e.g. 1996-04-01}
\item{status}{the status code, e.g. "M"
(\code{M} Member State, \code{C} Comprehensive Agreement State, \code{T} Transitional State)}
}
These are useful to grab the right spatial polygons in case of need.
Note: Kosovo is also included in the list.
}
\keyword{datasets}
