% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trajectories.R
\name{point_profile_tbl}
\alias{point_profile_tbl}
\title{Return a reference to the Point Profile table}
\usage{
point_profile_tbl(conn = NULL)
}
\arguments{
\item{conn}{Database connection or instantiate the default one.}
}
\value{
a \code{\link[dplyr:tbl]{dplyr::tbl()}} referencing the Oracle table for point profiles.
}
\description{
The returned \code{\link[dplyr:tbl]{dplyr::tbl()}} is referencing the point profiles table in PRISME.
You can use \code{dplyr}/\code{dbplyr} verbs to filter, join, ... with other
datasets.
}
\section{Note}{
You need to either provide a connection \code{conn} that has access to
\code{FSD.ALL_FT_POINT_PROFILE} or go with the default which uses
PRU_DEV to establish a \code{\link[=db_connection]{db_connection()}}.
}

\examples{
\dontrun{
pt <- point_profile_tbl()

# if you re-use DB connections
conn <- eurocontrol::db_connection("PRU_DEV")
pt <- point_profile_tbl(conn = conn)

# ... do something else with conn
# ...
# then manually close the connection to the DB
DBI::dbDisconnect(conn)
}
}
