% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iata_season.R
\name{season_iata}
\alias{season_iata}
\title{return the interval for an IATA season}
\usage{
season_iata(year, season = "summer")
}
\arguments{
\item{year}{the year for the season definition}

\item{season}{the (northern hemisphere) season, either "summer" (default) or "winter"}
}
\value{
an interval for the season definition, end/start dates are inclusive
}
\description{
IATA summer season begins on the last Sunday of March and ends on the last Saturday of October.
IATA winter season begins on the last Sunday of October and ends Saturday of before next year
summer season.
}
\examples{
\dontrun{
season_iata(2019)
}
}
