% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adrr.R
\name{adrr_flights_tidy}
\alias{adrr_flights_tidy}
\title{Extract ADRR flight list in an interval of dates}
\usage{
adrr_flights_tidy(conn = NULL, wef, til)
}
\arguments{
\item{conn}{Database connection or instantiate the default one.}

\item{wef}{\strong{W}ith \strong{EF}fect date (included) at Zulu time
in a format recognized by \code{\link[lubridate:as_date]{lubridate::as_datetime()}}}

\item{til}{un\strong{TIL}l date (excluded) at Zulu time
in a format recognized by \code{\link[lubridate:as_date]{lubridate::as_datetime()}}}
}
\value{
A \code{\link[dbplyr:tbl.src_dbi]{dbplyr::tbl_dbi()}} with the following columns (as defined in the
Aviation Data Repository for Research manual):
\subsection{Flight details}{
\itemize{
\item \verb{ECTRL ID}: flight unique id.
\item \code{ADEP}/\code{ADES}:
\item \verb{ADEP Latitude}/\verb{ADES Latitude}:
\item \verb{ADEP Longitude}/\verb{ADES Longitude}:
\item \verb{FILED OFF BLOCK TIME}:
\item \verb{FILED ARRIVAL TIME}:
\item \verb{ACTUAL OFF BLOCK TIME}:
\item \verb{ACTUAL ARRIVAL TIME}:
\item \verb{AC Type}:
\item \verb{AC Operator}:
\item \verb{AC Registration}:
\item \verb{ICAO Flight Type}:
\item \verb{STATFOR Market Segment}:
\item \verb{Requested FL}:
\item \verb{Actual Distance Flown (nm)}:
}
}
}
\description{
Extract ADRR flight list in an interval of dates
}
\section{Note}{
You need to either provide a connection \code{conn} that has access to
\code{SWH_FCT.DIM_FLIGHT_TYPE_RULE} (for \code{FLT_RULES}),
\code{PRUDEV.V_COVID_DIM_AO} (for aircraft and aircraft group info) and
\code{SWH_FCT.V_FAC_FLIGHT_MS} (for market segment info) or go with the default
which uses PRU_READ to establish a \code{\link[=db_connection]{db_connection()}}.
}

\examples{
\dontrun{
withr::local_envvar(c(TZ = "UTC", ORA_SDTZ = "UTC", NLS_LANG = ".AL32UTF8"))
conn <- withr::local_db_connection(db_connection(schema = "PRU_READ"))
wef <- "2025-01-01"
til <- "2025-01-04"
adrr_flights_tidy(conn, wef, til)
}
}
