% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{aircraft_type}
\alias{aircraft_type}
\title{ICAO's Aircraft types}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 2716 rows and 7 columns.
}
\usage{
aircraft_type
}
\description{
A data frame with the following fields
\describe{
\item{designator}{the aircraft type designator, e.g. "A310".}
\item{aircraft_description}{the aircraft description, e.g. "LandPlane".}
\item{description}{the aircraft , e.g. "IT".}
\item{wtc}{the aircraft \href{https://www.skybrary.aero/articles/icao-wake-turbulence-category}{wake turbulence category}, e.g. "M".}
\item{engine_count}{the number pf engines, e.g. "2".}
\item{engine_type}{the engine type, e.g. "Jet".}
\item{last_updated}{the last updated date when the data have been updated, e.g. "2023-05-19".}
}
}
\keyword{datasets}
