% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{airlines_tbl}
\alias{airlines_tbl}
\title{Return a reference to the Airlines table}
\usage{
airlines_tbl(conn = NULL)
}
\arguments{
\item{conn}{Database connection or instantiate the default one.}
}
\value{
a \code{\link[dplyr:tbl]{dplyr::tbl()}} referencing the Oracle table for airlines.
}
\description{
The returned \code{\link[dplyr:tbl]{dplyr::tbl()}} is referencing the airlines table in PRISME.
You can use \code{dplyr}/\code{dbplyr} verbs to filter, join, ... with other
datasets.
}
\section{Note}{
You need to either provide a connection \code{conn}
that has access to \code{PRU_DEV.V_COVID_DIM_AO} or go with the
default which uses PRU_DEV to establish a \code{\link[=db_connection]{db_connection()}}.
}

\examples{
\dontrun{
arl <- airlines_tbl()
# other operations on arl, i.e. filtering,
# followed by a collect() to retrieve the concrete data frame
arl_filtered <- arl |>
   dplyr::filter(AO_ISO_CTRY_CODE == "IT") |>
   collect()

# NOTE: you can reuse the connection for other API calls
arl$src$con

# IMPORTANT: at the end close the DB connection
DBI::dbDisconnect(arl$src$con)
}
}
