% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{airlines_tbl}
\alias{airlines_tbl}
\title{Return a reference to the Airlines table}
\usage{
airlines_tbl(con = NULL)
}
\arguments{
\item{con}{DB connection to be used, if NULL instantiate one.}
}
\value{
a tbl referencing the Oracle table for airlines
}
\description{
The returned \code{\link[dplyr:tbl]{dplyr::tbl()}} is referencing the airlines table in PRISME.
You can use \code{dplyr}/\code{dbplyr} verbs to filter, join, ... with other
datasets.

\strong{NOTE}: you need access to PRU_DEV schema
}
\examples{
\dontrun{
arl <- airlines_tbl()
}
}
