% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{airlines_tbl}
\alias{airlines_tbl}
\title{Return a reference to the Airlines table}
\usage{
airlines_tbl(con = NULL)
}
\arguments{
\item{con}{Database connection or instantiate the default one.}
}
\value{
a \link{tbl} referencing the Oracle table for airlines.
}
\description{
The returned \link{tbl} is referencing the airlines table in PRISME.
You can use \code{dplyr}/\code{dbplyr} verbs to filter, join, ... with other
datasets.
}
\section{Note}{
You need to either provide a connection \code{con}
that has access to \code{PRU_DEV.V_COVID_DIM_AO} or go with the
default which uses PRU_DEV to establish a \code{\link[=db_connection]{db_connection()}}.
}

\examples{
\dontrun{
arl <- airlines_tbl()
}
}
