% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{airlines_tbl}
\alias{airlines_tbl}
\title{Return a reference to the Airlines table}
\usage{
airlines_tbl()
}
\value{
a tbl referencing the Oracle table for airlines
}
\description{
The returned \code{tbl} is referencing the airline table in PRISME.
You can use \code{dplyr}/\code{dbplyr} verbs to filter, join, ... with other
datasets.
NOTE: you need access to PRU_DEV
}
\examples{
\dontrun{
arl <- airlines_tbl()
}
}
