% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/airlines.R
\name{airlines_tidy}
\alias{airlines_tidy}
\title{Airline info including group affiliation}
\usage{
airlines_tidy(conn = NULL)
}
\arguments{
\item{conn}{Optional connection to the PRU_DEV schema.}
}
\value{
A \code{\link[dplyr:tbl]{dplyr::tbl()}} with the following columns:
\itemize{
\item \code{AO_CODE}: the the \href{https://en.wikipedia.org/wiki/List_of_airline_codes}{ICAO Airline Designator}, i.e. 'OAL'
\item \code{AO_NAME}: the airline's name, i.e. 'Olympic'
\item \code{AO_GRP_CODE}: the airline's affiliation group code, i.e. 'AEE_GRP'
\item \code{AO_GRP_NAME}: the airline's affiliation group, i.e. 'AEGEAN Group'
\item \code{AO_ISO_CTRY_CODE}: the ISO2C code of the airline's country, i.e. 'GR'
\item \code{EU}: (a character) whether the airlines is in a EUROCONTROL's Member State
(full, comprehensive or transition plus Kosovo), i.e. 'TRUE'
}
}
\description{
Airline info including group affiliation
}
\section{Note}{
You need to either provide a connection \code{conn} that has access to \code{PRUDEV.V_COVID_DIM_AO}
or go with the default which uses PRU_DEV to establish a \code{\link[=db_connection]{db_connection()}}.
}

\examples{
\dontrun{
arls <- airlines_tidy()
# other operations on arls, i.e. filtering,
# followed by a collect() to retrieve the concrete data frame
arls_filtered <- arls |>
   filter(stringr::str_starts("A")) |>
   collect()

# NOTE: you can reuse the connection for other API calls
conn <- arls$src$con

# other ops requiring conn
# ...

# IMPORTANT: close the DB connection
DBI::dbDisconnect(conn)
}
}
