% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trajectories.R
\name{airspace_profile_tbl}
\alias{airspace_profile_tbl}
\title{Return a reference to the Airspace Profile table}
\usage{
airspace_profile_tbl(con = NULL)
}
\arguments{
\item{con}{Database connection or instantiate the default one.}
}
\value{
a \code{\link[dplyr:tbl]{dplyr::tbl()}} referencing the Oracle table for airspace profiles.
}
\description{
The returned \code{\link[dplyr:tbl]{dplyr::tbl()}} is referencing the airspace profiles table in PRISME.
You can use \code{dplyr}/\code{dbplyr} verbs to filter, join, ... with other
datasets.
}
\section{Note}{
You need to either provide a connection \code{con} that has access to \code{FSD.ALL_FT_ASP_PROFILE} or
go with the default which uses PRU_DEV to establish a \code{\link[=db_connection]{db_connection()}}.
}

\examples{
\dontrun{
prf <- airspace_profile_tbl()
}
}
