% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{airspace_profile_tbl}
\alias{airspace_profile_tbl}
\title{Return a reference to the Airspace Profile table}
\usage{
airspace_profile_tbl(profile = "CTFM")
}
\arguments{
\item{profile}{the \href{https://ansperformance.eu/definition/flight-models/}{model of the trajectory}
profile (default: 'CTFM'), one of:
\itemize{
\item 'FTFM'
\item 'RTFM'
\item 'CTFM'
\item 'CPF'
\item 'DCT'
\item 'SCR'
\item 'SRR'
\item 'SUR'
}}
}
\value{
a tbl referencing the Oracle table for airspace profiles
}
\description{
The returned \code{\link[dplyr:tbl]{dplyr::tbl()}} is referencing the flights table in PRISME.
You can use \code{dplyr}/\code{dbplyr} verbs to filter, join, ... with other
datasets.

\strong{NOTE}: you need access to PRU_DEV schema
}
\examples{
\dontrun{
prf <- airspace_profile_tbl(profile = "FTFM")
}
}
