% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trajectories.R
\name{airspace_profile_tbl}
\alias{airspace_profile_tbl}
\title{Return a reference to the Airspace Profile table}
\usage{
airspace_profile_tbl(conn = NULL)
}
\arguments{
\item{conn}{Database connection or instantiate the default one.}
}
\value{
a \code{\link[dbplyr:tbl.src_dbi]{dbplyr::tbl_dbi()}} referencing the Oracle table for airspace profiles.
}
\description{
The returned \code{\link[dbplyr:tbl.src_dbi]{dbplyr::tbl_dbi()}} is referencing the airspace profiles table in PRISME.
You can use \code{dplyr}/\code{dbplyr} verbs to filter, join, ... with other
datasets.
}
\section{Note}{
You need to either provide a connection \code{conn} that has access to \code{FSD.ALL_FT_ASP_PROFILE} or
go with the default which uses PRU_DEV to establish a \code{\link[=db_connection]{db_connection()}}.
}

\examples{
\dontrun{
pp <- airspace_profile_tbl()
# other operations on pp, i.e. filtering,
# followed by a collect() to retrieve the concrete data frame
# IMPORTANT: close the DB connection when done
DBI::dbDisconnect(pp$src$con)

# if you use a DB connection for many different APIs
conn <- eurocontrol::db_connection("PRU_DEV")
pp <- airspace_profile_tbl(conn = conn)

# ... do something else with conn
# ...
# then manually close the connection to the DB
DBI::dbDisconnect(conn)
}
}
