% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{db_connection}
\alias{db_connection}
\title{Provide a connection to the relevant Oracle database}
\usage{
db_connection(schema = "PRU_PROD")
}
\arguments{
\item{schema}{the Oracle DB schema to connect to.
Note: this is more the prefix of the environment variable
where the credentials are stored, like `<schema>_USR`,
`<schema>_PWD` and `<schema>_DBNAME`.
Possible values are `PRU_PROD`, `PRU_DEV`, `PRU_TEST`, ...}
}
\value{
DB connection
}
\description{
Provide a connection to the relevant Oracle database
}
\examples{
\dontrun{
con <- db_connection()
}
}
