% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{db_connection}
\alias{db_connection}
\title{Provide a connection to the relevant Oracle database}
\usage{
db_connection(schema = "PRU_PROD")
}
\arguments{
\item{schema}{the Oracle DB schema to connect to.}
}
\value{
A connection to a database (specifically an implementation of \link[DBI:DBIConnection-class]{DBI::DBIConnection}
for an Oracle database.)
}
\description{
Provide a connection to the relevant Oracle database
}
\section{Note}{
The \code{schema} is in fact the prefix of the environment variables
where the credentials are stored, like \verb{<schema>_USR},
\verb{<schema>_PWD} and \verb{<schema>_DBNAME}.
Possible values for \code{schema} are \code{PRU_PROD}, \code{PRU_DEV},
\code{PRU_TEST}, ...
}

\examples{
\dontrun{
conn <- db_connection()
# ... perform other API operations re-using the same connection
# ...
DBI::dbDisconnect(conn)
}
}
