% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{flights_airspace_profiles_tidy}
\alias{flights_airspace_profiles_tidy}
\title{Extract the flights list for the airspace profile segments intersecting an interval of interest}
\usage{
flights_airspace_profiles_tidy(
  con = NULL,
  wef,
  til,
  airspace = "FIR",
  profile = "CTFM"
)
}
\arguments{
\item{con}{DB connection to be used, if NULL instantiate one.}

\item{wef}{\strong{W}ith \strong{EF}fect date (included) at Zulu time
in a format recognized by \code{\link[lubridate:as_date]{lubridate::as_datetime()}}}

\item{til}{u\strong{TIL}l date (excluded) at Zulu time
in a format recognized by \code{\link[lubridate:as_date]{lubridate::as_datetime()}}}

\item{airspace}{the type of airspace (default: 'FIR'), one of:
\itemize{
\item 'FIR' (\href{https://observablehq.com/@openaviation/flight-information-regions}{Flight Information Region})
\item 'NAS' (National Airspace)
\item 'AUA' (ATC Unit Airspace)
\item 'ES' (Elementary Sector)
}}

\item{profile}{the \href{https://ansperformance.eu/definition/flight-models/}{model of the trajectory}
profile (default: 'CTFM'), one of:
\itemize{
\item 'FTFM'
\item 'RTFM'
\item 'CTFM'
\item 'CPF'
\item 'DCT'
\item 'SCR'
\item 'SRR'
\item 'SUR'
}}
}
\value{
a \link{tbl} with the same columns as \link{flights_tidy}
}
\description{
The returned \link{tbl} includes scheduled and non-scheduled flights whose airspace segments
temporally intersecting the right-opened interval \verb{[wef, til)}.
General aviation, State, military and sensitive flight are excluded.

\strong{NOTE}: you need access to PRU_DEV schema
}
\examples{
\dontrun{
asp_profs <- flights_airspace_profiles_tidy(wef = "2023-01-01", til = "2023-04-01")
}
}
