% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{flights_tbl}
\alias{flights_tbl}
\title{Return a reference to the Flights table}
\usage{
flights_tbl(conn = NULL)
}
\arguments{
\item{conn}{Database connection or instantiate the default one.}
}
\value{
a \code{\link[dplyr:tbl]{dplyr::tbl()}} referencing the Oracle table for flights.
}
\description{
The returned \code{\link[dplyr:tbl]{dplyr::tbl()}} is referencing the flights table in PRISME.
You can use \code{dplyr}/\code{dbplyr} verbs to filter, join, ... with other
datasets.
}
\section{Note}{
You need to either provide a connection \code{conn} that has access to \code{SWH_FCT.FAC_FLIGHT} or
go with the default which uses PRU_DEV to establish a \code{\link[=db_connection]{db_connection()}}.
}

\examples{
\dontrun{
flt <- flights_tbl()
# other operations on arl, i.e. filtering,
# followed by a collect() to retrieve the concrete data frame
flt_filtered <- flt |> dplyr::filter(AO_ISO_CTRY_CODE == "IT") |> collect()
# IMPORTANT: close the DB connection
DBI::dbDisconnect(flt$src$con)
}

}
