% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{flights_tbl}
\alias{flights_tbl}
\title{Return a reference to the Flights table}
\usage{
flights_tbl()
}
\value{
a tbl referencing the Oracle table for flight
}
\description{
The returned `tbl` is referencing the flight table in PRISME.
You can use `dplyr`/`dbplyr` verbs to filter, join, ... with other
datasets.
NOTE: you need access to PRU_DEV
}
\examples{
\dontrun{
arl <- flights_tbl()
}
}
