% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{flights_tbl}
\alias{flights_tbl}
\title{Return a reference to the Flights table}
\usage{
flights_tbl(con = NULL)
}
\arguments{
\item{con}{DB connection to be used, if NULL instantiate one.}
}
\value{
a tbl referencing the Oracle table for flight
}
\description{
The returned \code{\link[dplyr:tbl]{dplyr::tbl()}} is referencing the flights table in PRISME.
You can use \code{dplyr}/\code{dbplyr} verbs to filter, join, ... with other
datasets.

\strong{NOTE}: you need access to PRU_DEV schema
}
\examples{
\dontrun{
arl <- flights_tbl()
}

}
