% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flights.R
\name{flights_tidy}
\alias{flights_tidy}
\title{Extract a clean flights list in an interval}
\usage{
flights_tidy(
  conn = NULL,
  wef,
  til,
  icao_flt_types = c("S", "N"),
  ids = NULL,
  include_sensitive = FALSE,
  include_military = FALSE,
  include_head = FALSE
)
}
\arguments{
\item{conn}{Database connection or instantiate the default one.}

\item{wef}{\strong{W}ith \strong{EF}fect date (included) at Zulu time
in a format recognized by \code{\link[lubridate:as_date]{lubridate::as_datetime()}}}

\item{til}{un\strong{TIL}l date (excluded) at Zulu time
in a format recognized by \code{\link[lubridate:as_date]{lubridate::as_datetime()}}}

\item{icao_flt_types}{the types of flights as described below in
\code{ICAO_FLT_TYPE}, default \code{c('S', 'N')}, NULL includes all
notwithstanding other argument options.
When including military via \code{include_military} you should
either pass NULL or make sure 'M' is included}

\item{ids}{list of \code{ID}s (aka \verb{SAM ID}) to return, default NULL for all flights}

\item{include_sensitive}{include sensitive flights, default FALSE}

\item{include_military}{include military flights, default FALSE}

\item{include_head}{include Head of State flights, default FALSE}
}
\value{
A \code{\link[dbplyr:tbl.src_dbi]{dbplyr::tbl_dbi()}} with the following columns (grouped here by
flight details, aerodrome details, aircraft info, aircraft operator
info and operational details):
\subsection{Flight details}{
\itemize{
\item FLT_UID: flight unique id.
\item ID: the so called \verb{SAM ID}, used internally by PRISME
\item AIRCRAFT_ID: the \href{https://www.skybrary.aero/articles/aircraft-call-sign}{callsign}
of the relevant flight, e.g. BAW6VB.
\item LOBT: \strong{L}ast received \strong{O}ff-\strong{B}lock \strong{T}ime.
\item IOBT: \strong{I}nitial \strong{O}ff-\strong{B}lock \strong{T}ime.
\item FLT_RULES (see \href{https://www.skybrary.aero/articles/flight-plan-completion}{FPL Item 8}):
which sets of regulations the flight is operated under.
Possible values are:
\itemize{
\item \code{I} for IFR
\item \code{V} for VFR
\item \code{Y} first IFR thereafter VFR
\item \code{Z} first VFR thereafter IFR
}
\item ICAO_FLT_TYPE (see \href{https://www.skybrary.aero/articles/flight-plan-completion}{FPL Item 8}):
flight type. Possible values:
\itemize{
\item \code{S} for scheduled air service
\item \code{N} for non-scheduled air service
\item \code{G} for general aviation
\item \code{M} for military (note: filtered out)
\item \code{X} for other than the preceding categories
}
\item RULE_NAME: market segment type as defined on the
\href{https://www.eurocontrol.int/publication/market-segment-rules}{Market Segment Rules},
it can be:
\itemize{
\item “Mainline”
\item “Regional”
\item “Low-Cost”
\item “Business Aviation”
\item “All-Cargo”
\item “Charter” (Non-Scheduled)
\item “Military”
\item “Other”
\item "Not classified"
}
\item SENSITIVE: 'Y' if sensitive
\item SPECIAL_EXEMPT: reasons for special handling by ATS.
One of:
\itemize{
\item "AEAP" ATFM exemption approved
\item "EMER" emergency
\item "FIRE" fire fighting
\item "HEAD" flights with Head of State status
\item "MEDE" medical evacuation
\item "NEXE" not exempted
\item "SERE" search & rescue
}
}
}

\subsection{Aerodrome details}{
\itemize{
\item ADEP: \href{https://observablehq.com/@openaviation/airports}{ICAO code} of the
\strong{A}erodrome of \strong{DEP}arture
\item NAME_ADEP: the (AIU) name of the \code{ADEP} airport
\item COUNTRY_CODE_ADEP: the ISO 2-alpha country code for \code{ADEP}
\item COUNTRY_NAME_ADEP: the country name for \code{ADEP}
\item ADES: \href{https://observablehq.com/@openaviation/airports}{ICAO code} of the
\strong{A}erodrome of \strong{DES}tination (different from \code{ADES_FILED} in case of
diversion)
\item NAME_ADES: the (AIU) name of the \code{ADES} airport
\item COUNTRY_CODE_ADES: the ISO 2-alpha country code for \code{ADES}
\item COUNTRY_NAME_ADES: the country name for \code{ADES}
\item ADES_FILED: \href{https://observablehq.com/@openaviation/airports}{ICAO code} of
the \strong{A}erodrome of \strong{DES}tination filed in the Flight Plan.
\strong{Note}: it can be different from \code{ADES} in case of diversion
\item NAME_ADES_FILED: the (AIU) name of the \code{ADES_FILED} airport
\item COUNTRY_CODE_ADES_FILED: the ISO 2-alpha country code for \code{ADES_FILED}
\item COUNTRY_NAME_ADES_FILED: the country name for \code{ADES_FILED}
}
}

\subsection{Aircraft info}{
\itemize{
\item REGISTRATION: the \href{https://en.wikipedia.org/wiki/Aircraft_registration}{aircraft registration}
(with spaces, dashes, ... stripped), e.g. GEUUU.
\item AIRCRAFT_ADDRESS: the \href{https://en.wikipedia.org/wiki/Aviation_transponder_interrogation_modes#ICAO_24-bit_address}{ICAO 24-bit address}
of the airframe for ADS-B/Mode S broadcasting.
\item AIRCRAFT_TYPE_ICAO_ID: the \href{https://www.icao.int/publications/doc8643/pages/search.aspx}{ICAO code for the aircraft type}, for example A30B for an
Airbus A-300B2-200.
\item WK_TBL_CAT (see \href{https://www.skybrary.aero/articles/flight-plan-completion}{FPL Item 9}): wake turbulence category, can be
\itemize{
\item \code{L} LIGHT, i.e. maximum certificated takeoff mass of 7000 kg (15_500 lbs) or less.
\item \code{M} MEDIUM, i.e maximum certificated takeoff mass less than 136_000 kg (300_000 lbs),
but more than 7_000 kg (15_500 lbs)
\item \code{H} HEAVY, i.e. maximum certificated takeoff mass of 136_000 kg (300_000 lbs) or more
(except those specified as \code{J})
\item \code{J} SUPER, presently the only the AIRBUS A-380-800
}
}
}

\subsection{Aircraft operator details}{
\itemize{
\item AIRCRAFT_OPERATOR: the \href{https://en.wikipedia.org/wiki/List_of_airline_codes}{ICAO Airline Designator},
i.e. \code{OAL} for \code{Olympic}
\item AO_GRP_CODE: Aircraft Operator group (code), i.e. AEE_GRP
\item AO_GRP_NAME: : Aircraft Operator group (name), i.e. AEGEAN Group
\item AO_ISO_CTRY_CODE: ISO country code for AO
}
}

\subsection{Operational details}{
\itemize{
\item EOBT_1: \strong{E}stimated \strong{O}ff-\strong{B}lock \strong{T}ime for FPL-based (M1) trajectory
\item ARVT_1: \strong{AR}ri\strong{V}al \strong{T}ime for FPL-based (M1) trajectory
\item TAXI_TIME_1: Taxi time for FPL-based (M1) trajectory
\item AOBT_3: \strong{A}ctual \strong{O}ff-\strong{B}lock \strong{T}ime for flown (M3) trajectory
\item ARVT_3: \strong{ARV}ival \strong{T}ime for flown (M3) trajectory
\item TAXI_TIME_3: Taxi time for flown (M3) trajectory
\item RTE_LEN_1: route length (in Nautical Miles) for FPL-based (M1) trajectory
\item RTE_LEN_3: route length (in Nautical Miles) for for flown (M3) trajectory
\item FLT_DUR_1: route duration (in minutes) for FPL-based (M1) trajectory
\item FLT_DUR_3: route length (in minutes) for flown (M3) trajectory
\item FLT_TOW: takeoff weight
}
}
}
\description{
The returned \code{\link[dbplyr:tbl.src_dbi]{dbplyr::tbl_dbi()}} includes scheduled and non-scheduled flight
departing in the right-opened interval \verb{[wef, til)}.

Defaults values will assure that General aviation, State, military and
sensitive flight will excluded.
They can be retrieved via the other function call arguments in case of need.
}
\section{Note}{
You need to either provide a connection \code{conn} that has access to
\code{SWH_FCT.DIM_FLIGHT_TYPE_RULE} (for \code{FLT_RULES}),
\code{PRUDEV.V_COVID_DIM_AO} (for aircraft and aircraft group info) and
\code{SWH_FCT.V_FAC_FLIGHT_MS} (for market segment info) or go with the default
which uses PRU_DEV to establish a \code{\link[=db_connection]{db_connection()}}.
}

\examples{
\dontrun{
flts <- flights_tidy(wef = "2023-01-01", til = "2023-01-05")
# other operations on flts, i.e. filtering,
# followed by a collect() to retrieve the concrete data frame
flts_filtered <- flts |>
   filter(TO_DATE("2023-06-01 10:00", "YYYY-MM-DD HH24:MI") <= IOBT,
                  IOBT < TO_DATE("2023-01-02 10:30", "YYYY-MM-DD HH24:MI")) |>
   collect()

# NOTE: you can reuse the connection for other API calls
conn <- flts$src$con

# other ops requiring conn
# ...

# IMPORTANT: close the DB connection
DBI::dbDisconnect(conn)
}
}
