% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trajectories.R
\name{point_profiles_tidy}
\alias{point_profiles_tidy}
\title{Export point profile from NM trajectories}
\usage{
point_profiles_tidy(con = NULL, wef, til, model = "CTFM", bbox = NULL)
}
\arguments{
\item{con}{Database connection or instantiate the default one.}

\item{wef}{\strong{W}ith \strong{EF}fect date (included) at Zulu time
in a format recognized by \code{\link[lubridate:as_date]{lubridate::as_datetime()}}}

\item{til}{un\strong{TIL}l date (excluded) at Zulu time
in a format recognized by \code{\link[lubridate:as_date]{lubridate::as_datetime()}}}

\item{model}{the model of the profile: one of
CPF, CTFM, DCT, FTFM, SCR, SRR, SUR
[default: "CFTM"].}

\item{bbox}{(Optional) axis aligned bounding box
(xmin, ymin, xmax, ymax)}
}
\value{
a dataframe with trajectory data
}
\description{
Extract NM point profile trajectories from PRISME database
}
\section{Note}{
You need to either provide a connection \code{con} that has access to as noted in
\code{\link[=airspace_profile_tbl]{airspace_profile_tbl()}} and \code{\link[=flights_tidy]{flights_tidy()}} or go with the default which uses
PRU_DEV to establish a \code{\link[=db_connection]{db_connection()}}.
}

\examples{
\dontrun{
# export 1 day worth of NM (planned) trajectories
export_model_trajectory("2019-07-14", "2019-07-15", model = "FTFM")

# export 2 hours of NM (flown) trajectories
export_model_trajectory("2019-07-14 22:00", "2019-07-15")

# export 1 day of NM (flown) trajectories
point_profiles_tidy("2019-07-14", "2019-07-15", lobt_buffer = c(before = 24, after = 1.25))

# export all CTFM trajectories within a bounding box 40 NM around EDDF
bb <- c(xmin = 7.536746, xmax = 9.604390, ymin = 49.36732, ymax = 50.69920)
point_profiles_tidy("2019-01-01 00:00", "2019-01-02 00:00", bbox = bb)
}
}
\concept{read/export}
